Simplify the .htaccess analyzer
